Switch to global banner system
Per-board banners will be ignored in favor of global banners, which are
pulled from static/global-banner. Images can be paired with .txt files
that are used as tooltips on the global banners page.

Might be nice someday to switch a db-based method instead of this, but
this should work fine for now.